Move ExpressionAndValuePointers to top level
