+++ core/trino-main/src/test/java/io/trino/sql/planner/assertions/PatternRecognitionExpressionRewriter.java
@@ -18,8 +18,8 @@ import io.trino.spi.type.Type;
 import io.trino.sql.parser.SqlParser;
 import io.trino.sql.planner.Symbol;
 import io.trino.sql.planner.SymbolAllocator;
+import io.trino.sql.planner.rowpattern.ExpressionAndValuePointers;
 import io.trino.sql.planner.rowpattern.LogicalIndexExtractor;
-import io.trino.sql.planner.rowpattern.LogicalIndexExtractor.ExpressionAndValuePointers;
 import io.trino.sql.planner.rowpattern.ir.IrLabel;
 import io.trino.sql.tree.DereferenceExpression;
 import io.trino.sql.tree.Expression;
